Added register class

+++ Simulator/IRegister.cs
@@ -4,7 +4,7 @@ namespace Simulator
 {
     public interface IRegister
     {
-        int Value { get; set; }
+        byte Value { get; set; }
         string BinarytValue { get; }
 
         void SetBit(int bit, bool value);
